+++ Semana 2/node/src/routes/appointments.routes.ts
@@ -3,8 +3,11 @@ import { getCustomRepository } from "typeorm";
 import AppointmentsRepository from "../repositories/AppointmentsRepository";
 import { getParsedHour } from "../model/Appointments";
 import CreateAppointmentService from "../services/CreateAppointmentService";
+import ensureAuthenticated from '../middlewares/ensureAuthenticated'
 
 const appoinmentRouter = Router();
+
+appoinmentRouter.use(ensureAuthenticated)
 
 //let appointmentsRepository = new AppointmentsRepository();
 
@@ -17,7 +20,7 @@ appoinmentRouter.post("/", async (request, response) => {
 
     let appointment = await createService.execute({ parsedDate, provider_id });
 
-    return response.json(appointment);
+    return response.json({appointment});
   } catch (error) {
     return response.status(400).json({ error: error.message });
   }
@@ -27,6 +30,7 @@ appoinmentRouter.get("/", async (request, response) => {
   let appointmentsRepository = getCustomRepository(AppointmentsRepository);
 
   let appointments = await appointmentsRepository.find();
+  //eturn response.json({appointments,user:request.user}); //Teste da passagem de valores atraves do middleware
   return response.json(appointments);
 });