+++ Semana 2/node/src/routes/sessions.routes.ts
@@ -1,0 +1,25 @@
+import { Router } from "express";
+import AuthenticateUserService from "../services/AuthenticateUserService";
+
+const sessionsRouter = Router();
+
+sessionsRouter.post("/", async (request, response) => {
+  try {
+    let { email, password } = request.body;
+    const authService = new AuthenticateUserService();
+
+    let {user,token} = await authService.execute({
+      email,
+      password,
+    });
+
+
+    delete user.password
+
+    return response.json({user,token})
+  } catch (err) {
+    return response.status(400).json({ error: err.message });
+  }
+});
+
+export default sessionsRouter;